Move Interface to separate module
